Commit message to 'google_maps_api.xml': Tried to fix the issue of not seeing the map in the emulator by using a different API key (this method did not work, so I reverted back to the first API key)
Commit message to 'HomePageActivity.java':
Added zoom functionality to the zoom button and placing a marker on a point the user taps. Please keep in mind, the 'Sat', 'Mark' and search bars do not have functionality at the present moment, It will not be an issue to add functionality to these elements though.

Commit message to 'activity_home_page2.xml':
Added buttons and a search bar.

Commit message to 'AndroidManifest.xml'
Added a permission.